Fall back to `application/octet-stream` if no mimetype is set
This allows browsers to still display the image if it is in an `<img>`
tag (at least Firefox does it) even though it doesn't have an image
content-type.

This happens currently when a file has no extension that indicates it's
type, as mime-type is not guessed from the content for now.