Move max_size handling out of storage, add mimetype support
This implements the following changes based on lessons from
implementing our first real cloud backend (GCP):

1. Let's only care about max_size in the CKAN layer. If we can't figure
out the file size before we upload, let's just ignore it. This really
simplifies the storage backend logic. For uploaded streams the stream
may not be seekable, but we do get 'Content-length' headers from the
client. However, making storage backends aware of this HTTP-specific
semantics is wrong, so let's keep max_size handling up the stack.

2. Allow uploader to pass mime type to storage when uploading, if known.
Again this is done because there are HTTP-specific features that support
guessing the mime type and it doesn't make sense to implement the
fallback logic in each backend.

The local storage backend doesn't store mime type when uploading, but
guesses it when downloading.